x
Module lint

Structs

Args

Statics

CREATE_DAPP_TEMPLATE_DIR 🔒

EXTERNAL_CRATE_DIR 🔒

LICENSE_HEADER 🔒

Functions

handle_lint_results_exclude_external_crate_checks

Define custom handler so we can skip certain lints on certain files. This is a temporary till we upstream this logic

run